requirements.txt
p
pytesseract
Pillow
opencv-python
psycopg2